- Implement AwsConfig Configuration class for loading aws credentials from application.properties and initializing Sqs client. - Add getChapter and updateChapter methods to ChapterService - Add aws sqs client dependency - Implement SqsProducerService to allow enqueuing messages to the TextToSummarize queue. - Implement SummaryMessageListener to listen for generated Summaries on application startup and update database relations as necessary for videos and chapters - Add getVideo and updateVideo methods to ChapterService - Refactor VideoService to use sqs instead of HTTP for generating summaries

+++ backend/src/main/java/com/concise/backend/VideoRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 import java.util.List;
 
-public interface VideoRepository extends JpaRepository<VideoEntity, Integer> {
+public interface VideoRepository extends JpaRepository<VideoEntity, Long> {
     List<VideoEntity> findAllByYoutubeIdAndSummaryLanguageAndUserId(String youtubeId, String language, long userId);
     List<VideoEntity> findAllByUserIdOrderByIdDesc(Long userId);
     Optional<VideoEntity> findByIdAndUserId(Long id, Long userId);
